Revert "Refactoring connector"
