+++ Connector/package.json
@@ -3,11 +3,16 @@
     "version": "1.0.12",
     "description": "iModel Connector to push Sample Data to iModelHub",
     "scripts": {
-        "build": "tsc && mkdirp lib/Schemas && cp Schemas/*.xml  lib/schemas/ && mkdirp lib/Connector/src/test/assets && cp Connector/src/test/assets/*.db  lib/Connector/src/test/assets/ ",
+        "copyFiles": "npx babel src --out-dir lib --copy-files",
+        "pretest": "cpx ./src/test/logging.config.json ./lib/test",
+        "test": "npm run test:unit ",
+        "test:unit": "mocha --opts ./src/test/unit/mocha.opts \"./src/test/unit/**/*.test.ts*\"",
+        "test:integration": "npm run pretest && betools test --testDir=\"./lib/test/integration\"",
+        "build": "tsc && npm run copyFiles",
         "clean": "rimraf lib",
-        "lint": "eslint ./Connector/src/**/*.ts 1>&2",
+        "lint": "eslint ./src/**/*.ts 1>&2",
         "blint": "npm run build && npm run lint",
-        "start": "cd Connector/Extractor/extractor/ && RunExtractor.cmd --clean && RunExtractor.cmd --all && RunExtractor.cmd --copy && cd ../../../ && npm run clean && npm run build && node lib/Main.js",
+        "start": "cd ../Extractor/extractor/ && RunExtractor.cmd --clean && RunExtractor.cmd --all && RunExtractor.cmd --copy && cd ../../Connector/ && npm run clean && npm run build && node lib/main.js",
         "insert": "node lib/insertElement.js",
         "purge": "node lib/PurgeBriefcase.js"
     },
