+++ Connector/src/Connector.ts
@@ -42,9 +42,9 @@ export class Connector extends IModelBridge {
 
   public async importDomainSchema(_requestContext: AuthorizedClientRequestContext | ClientRequestContext): Promise<any> {
     if (this.sourceDataState === ItemState.New) {
-      const functionalSchemaPath = path.join(__dirname, "../../schemas/Functional.ecschema.xml");
-      const spatialCompositionSchemaPath = path.join(__dirname, "../../schemas/SpatialComposition.ecschema.xml");
-      const buildingSpatialSchemaPath = path.join(__dirname, "../../schemas/BuildingSpatial.ecschema.xml");
+      const functionalSchemaPath = path.join(__dirname, "./schema/Functional.ecschema.xml");
+      const spatialCompositionSchemaPath = path.join(__dirname, "./schema/SpatialComposition.ecschema.xml");
+      const buildingSpatialSchemaPath = path.join(__dirname, "./schema/BuildingSpatial.ecschema.xml");
       await this.synchronizer.imodel.importSchemas(_requestContext, [functionalSchemaPath, spatialCompositionSchemaPath, buildingSpatialSchemaPath]);
     }
   }
