+++ Connector/src/schema/BuildingSpatial.ecschema.xml
