+++ Connector/src/schema/IoT.ecschema.xml
