+++ Connector/src/schema/SpatialComposition.ecschema.xml
