+++ Extractor/extractor/RunExtractor.cmd
@@ -13,7 +13,7 @@ goto usage
 	IF NOT DEFINED pythonCmd Set pythonCmd=python.exe
 	echo Creating the intermediary .db files...
 	IF NOT EXIST output mkdir output
-	%pythonCmd% extractor.py ..\..\..\input/samplesheet.xlsx output/samplesheet.db # create
+	%pythonCmd% extractor.py input/samplesheet.xlsx output/samplesheet.db # create
 	@REM %pythonCmd% extractor.py input/SampleSheetV2.xlsx output/intermediary_v2.db # data change
 	@REM %pythonCmd% extractor.py input/SampleSheetV3.xlsx output/intermediary_v3.db # schema change (addition)
 	@REM %pythonCmd% extractor.py input/SampleSheetV4.xlsx output/intermediary_v4.db # schema change (deletion)
@@ -23,13 +23,17 @@ goto end
 :clean
   echo Running Extractor now...
 	rmdir /s /q output
-	del ..\..\src\test\assets\*.db
+	del ..\..\Connector\src\assets\*.db
+	del ..\..\Connector\src\test\assets\*.db
 goto end
 
 :copyToConnector
-IF NOT EXIST ..\..\src\test\assets mkdir ..\..\src\test\assets
-IF NOT EXIST ..\..\src\test\output mkdir ..\..\src\test\output
-copy .\output\*.db  ..\..\src\test\assets\
+IF NOT EXIST ..\..\Connector\src\assets mkdir ..\..\Connector\src\assets
+IF NOT EXIST ..\..\Connector\src\output mkdir ..\..\Connector\src\output
+IF NOT EXIST ..\..\Connector\src\test\assets mkdir ..\..\Connector\src\test\assets
+IF NOT EXIST ..\..\Connector\src\test\output mkdir ..\..\Connector\src\test\output
+copy .\output\*.db  ..\..\Connector\src\assets\ 
+copy .\output\*.db  ..\..\Connector\src\test\assets\
 echo Created the intermediary .db files and copied to Connector.
 goto end
